% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Sum two numeric objects.}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A natural number}
}
\value{
A list of all the numbers leading up to n,
replacing numbers divisible by 3 with Fizz, numbers divisible by 5 with Buzz,
and numbers divisible by both 3 and 5 with Fizz-Buzz
}
\description{
Because this is a very simple function, my explanation is short. These
paragraphs should explain everything you need to know.
}
\details{
This is still in the description part of the documentation and and it
will be until we see something that indicates a new section.
}
\examples{
FizzBuzz(4)
FizzBuzz(10)
}
