% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of daily maximum temperatures in Pulliam Airport.}
\format{
a data frame with 25449 observations
\describe{
  \item{DATE}{The date of observation.}
  \item{PRCP}{Daily maximum temperature in degrees Farhenheit.}
  \item{SNOW}{The amount of snowfall in inches.}
  \item{TMAX}{Daily maximum temperature in degrees Farhenheit.}
  \item{TMIN}{Daily maximum temperature in degrees Farhenheit.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily maximum temperatures in Pulliam Airport.
}
\keyword{datasets}
